\name{comp}
\alias{comp}
\title{Arranging information from GRanges produced by readsamples() on
per position (nucleotide) basis.}
\usage{
comp(euc_GR, cutoff = 1, fasta_file)
}
\arguments{
\item{euc_GR}{GRanges generated by readsamples() function}

\item{cutoff}{specifies cutoff length, only inserts of this length or longer
will be used for processing (default: 1)}

\item{fasta_file}{path to fasta file to which reads were mapped. Used to
report nucleotide at each position (not required)}
}
\value{
GRanges object with: 1) seqnames (RNAid), 2) start (position within
RNA), and metadata: 3) TCR (termination coverage ratio), 4) TC (termination
count), 5) Cover (coverage) and 6) PC (priming count) for each position
within each RNA.
}
\description{
comp() takes as input euc_GR GRanges object produced by readsamples()
and produces Comp_GR GRanges.
}
\examples{
dummy_euc_GR <- GRanges(seqnames="DummyRNA",
                        IRanges(start=round(runif(100)*100),
                        width=round(runif(100)*100+1)), strand="+",
                        EUC=round(runif(100)*100))
comp(dummy_euc_GR)
}
\author{
Lukasz Jan Kielpinski, Nikos Sidiropoulos
}
\references{
Kielpinski, L.J., and Vinther, J. (2014). Massive
parallel-sequencing-based hydroxyl radical probing of RNA accessibility.
Nucleic Acids Res.
}
\seealso{
\code{\link{readsamples}}, \code{\link{dtcr}},
\code{\link{slograt}}, \code{\link{swinsor}}, \code{\link{compdata}},
\code{\link{comp}}
}

