% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correct_oversaturation}
\alias{correct_oversaturation}
\title{Correcting EUC of oversaturated fragments.}
\usage{
correct_oversaturation(euc_GR, read_counts_file)
}
\arguments{
\item{euc_GR}{GRanges produced by readsamples() function}

\item{read_counts_file}{path to a file with observed read counts.}
}
\description{
If for a given fragment the number of observed unique barcodes is equal to the total barcode complexity (all combinations of barcodes are associated with a given fragment), then the readsamples function assignes infinite EUC.
This can be corrected by the function correct_oversaturation(). By comparing observed read counts with EUCs for other fragments it calculates the correction factor.
Then, for the oversaturated fragments it multiplies the observed read counts by the correction factor to estimate EUC. The assumption behind this correction is that fragments have similar rate of PCR duplicates production.
}

