\name{dtcr}
\alias{dtcr}
\title{Calculate deltaTCR.}
\usage{
dtcr(control_GR, treated_GR, window_size = 3, nt_offset = 1,
  bring_to_zero = TRUE, add_to)
}
\arguments{
\item{control_GR}{GRanges object made by comp() function from the control
sample.}

\item{treated_GR}{GRanges object made by comp() function from the treated
sample.}

\item{window_size}{if smoothing is to be performed, what should be the
window size?  (use only odd numbers to ensure that windows are centred on a
nucleotide of interest) (default: 3)}

\item{nt_offset}{how many nucleotides before a modification the reverse
transcription terminates. E.g. for HRF-Seq nt_offset=1 (default: 1)}

\item{bring_to_zero}{should in deltaTCR calculations negative deltaTCR's be
brought to 0 as was done in HRF-Seq paper (default: T)}

\item{add_to}{GRanges object made by other normalization function (dtcr(),
slograt(), swinsor(), compdata()) to which normalized values should be
added.}
}
\value{
GRanges object with "dtcr" (deltaTCR) and "dtcr.p" (p.value of
comparing control and treated calcualted with pooled two-proportion Z-test)
metadata.
}
\description{
Performs deltaTCR (dtcr) normalization given control and treated GRanges
generated by comp() function.
}
\examples{
dummy_euc_GR_control <- GRanges(seqnames="DummyRNA",
                                IRanges(start=round(runif(100)*100),
                                width=round(runif(100)*100+1)), strand="+",
                                EUC=round(runif(100)*100))
dummy_euc_GR_treated <- GRanges(seqnames="DummyRNA",
                                IRanges(start=round(runif(100)*100),
                                width=round(runif(100)*100+1)), strand="+",
                                EUC=round(runif(100)*100))
dummy_comp_GR_control <- comp(dummy_euc_GR_control)
dummy_comp_GR_treated <- comp(dummy_euc_GR_treated)
dtcr(control_GR=dummy_comp_GR_control, treated_GR=dummy_comp_GR_treated)
}
\author{
Lukasz Jan Kielpinski, Nikos Sidiropoulos
}
\references{
Kielpinski, L.J., and Vinther, J. (2014). Massive
parallel-sequencing-based hydroxyl radical probing of RNA accessibility.
Nucleic Acids Res.
}
\seealso{
\code{\link{comp}}, \code{\link{slograt}}, \code{\link{swinsor}},
\code{\link{compdata}}, \code{\link{GR2norm_df}}, \code{\link{plotRNA}},
\code{\link{norm2bedgraph}}
}

