\name{slograt}
\alias{slograt}
\title{Smooth Log2-ratio}
\usage{
slograt(control_GR, treated_GR, window_size = 5, nt_offset = 1,
  depth_correction = "all", pseudocount = 5, add_to)
}
\arguments{
\item{control_GR}{GRanges object made by comp() function from the control
sample.}

\item{treated_GR}{GRanges object made by comp() function from the treated
sample.}

\item{window_size}{if smoothing is to be performed, then what should be the
window size?  (use only odd numbers to ensure that windows are centred on a
nucleotide of interest) (default: 5)}

\item{nt_offset}{How many position in the 5' direction should the signal be
offset to account for the fact that reverse transcription termination occurs
before site of modification.}

\item{depth_correction}{One of three values: "no" - counts are used as
given, "all" - counts from sample with higher total sum of EUCs are
multiplied by sum of EUCs from sample with lower total sum of EUCs and
divided by sum of EUCs from sample with higher EUC count (default),
"RNA" as in "all" but on per RNA basis}

\item{pseudocount}{What pseudocount should be added to each nucleotide prior
to calculating log2 ratio (default: 5)}

\item{add_to}{GRanges object made by other normalization function (dtcr(),
slograt(), swinsor(), compdata()) to which normalized values should be
added.}
}
\value{
GRanges object with "slograt" (smooth log2 ratio) and "slograt.p"
(p.value of comparing control and treated) metadata.
}
\description{
Performs smooth-log2-ratio calculation given control and treated
GRanges generated by comp() function.
}
\examples{
dummy_euc_GR_control <- GRanges(seqnames="DummyRNA",
                                IRanges(start=round(runif(100)*100),
                                width=round(runif(100)*100+1)), strand="+",
                                EUC=round(runif(100)*100))
dummy_euc_GR_treated <- GRanges(seqnames="DummyRNA",
                                IRanges(start=round(runif(100)*100),
                                width=round(runif(100)*100+1)), strand="+",
                                EUC=round(runif(100)*100))
dummy_comp_GR_control <- comp(dummy_euc_GR_control)
dummy_comp_GR_treated <- comp(dummy_euc_GR_treated)
slograt(control_GR=dummy_comp_GR_control, treated_GR=dummy_comp_GR_treated)
}
\author{
Lukasz Jan Kielpinski, Nikos Sidiropoulos
}
\references{
Wan, Y., Qu, K., Zhang, Q.C., Flynn, R.A., Manor, O., Ouyang,
Z., Zhang, J., Spitale, R.C., Snyder, M.P., Segal, E., et al. (2014).
Landscape and variation of RNA secondary structure across the human
transcriptome. Nature 505, 706-709.
}
\seealso{
\code{\link{comp}}, \code{\link{dtcr}}, \code{\link{compdata}},
\code{\link{swinsor}}, \code{\link{GR2norm_df}}, \code{\link{plotRNA}},
\code{\link{norm2bedgraph}}
}

